requirements.txt
amqp==1.4.8
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.19
Django==1.9
django-bootstrap-toolkit==2.15.0
django-celery==3.1.17
django-decorators==0.1.1
djangorestframework==3.3.1
enum34==1.1.1
idna==2.0
ipaddress==1.0.15
kombu==3.0.30
MySQL-python==1.2.5
pyasn1==0.1.9
pycparser==2.14
pytz==2015.7
six==1.10.0
wheel==0.24.0
